+++ QA/M4PL/M4PL.Entities/Job/Job.cs
@@ -492,7 +492,7 @@ namespace M4PL.Entities.Job
         public string JobSignText { get; set; }
         public string JobSignLatitude { get; set; }
         public string JobSignLongitude { get; set; }
-        public decimal? JobQtyOrdered { get; set; }
+        public int? JobQtyOrdered { get; set; }
         public int? JobQtyActual { get; set; }
         public int? JobQtyUnitTypeId { get; set; }
         public string JobQtyUnitTypeIdName { get; set; }
